+++ src/backend/resolvers/mutations/usState/index.js
@@ -1,0 +1,9 @@
+const createUsState = require('./create')
+const updateUsState = require('./update')
+const deleteUsState = require('./delete')
+
+module.exports = {
+  createUsState,
+  updateUsState,
+  deleteUsState,
+}