+++ src/backend/typeDefs/mutations/index.js
@@ -34,6 +34,8 @@ const person = require('./person')
 
 const endUserTerms = require('./endUserTerms')
 
+const usState = require('./usState')
+
 const mutationType = gql`
   type Mutation {
     createUser(input: CreateUserInput!): CreateUserPayload
@@ -307,6 +309,10 @@ const mutationType = gql`
     updateEndUserTerms(
       input: UpdateEndUserTermsInput!
     ): UpdateEndUserTermsPayload
+
+    createUsState(input: CreateUsStateInput!): CreateUsStatePayload!
+    updateUsState(input: UpdateUsStateInput!): UpdateUsStatePayload!
+    deleteUsState(input: DeleteUsStateInput!): DeleteUsStatePayload!
   }
 `
 
@@ -344,4 +350,5 @@ module.exports = [
   importWorkbook,
   ...person,
   ...endUserTerms,
+  ...usState,
 ]