+++ src/frontend/api/queries.js
@@ -1,5 +1,21 @@
 import gql from 'graphql-tag'
 
+export const GET_US_STATES = gql`
+  query getUsStates {
+    usStates {
+      _id
+      state
+      stateLong
+      status
+      booksImpacted
+      law
+      lawLink
+      bill
+      surveyCommercialLivesPercentInsured
+    }
+  }
+`
+
 export const GET_OPEN_PAYMENTS = gql`
   query getOpenPayments($physicianProfileId: Float) {
     openPayments(physicianProfileId: $physicianProfileId) {