+++ src/frontend/hooks/useBom/boid-query-map.js
@@ -7,8 +7,9 @@ import {
   GET_OBM_SERVICES,
   GET_OBM_SERVICES_CATEGORIES,
   GET_PEOPLE,
+  GET_US_STATES,
   GET_THERAPEUTIC_AREAS,
-} from './../../api/queries'
+} from 'frontend/api/queries'
 
 export default {
   '5eac3251ac8a01743081f28d': GET_PATHWAYS_ORGANIZATIONS,
@@ -16,5 +17,6 @@ export default {
   '5ed81e5fb8ebf33703463750': GET_OBM_SERVICES,
   '5ed81ed8bcfdf6381562c17e': GET_OBM_SERVICES_CATEGORIES,
   '5eea22d5adbf920fa4320487': GET_PEOPLE,
+  '5f29d373b632504434da0f90': GET_US_STATES,
   '5f2ac32600b9d74431f9bc6f': GET_THERAPEUTIC_AREAS,
 }